api.py/requirements.txt
pybase64
requests
python-wordpress-xmlrpc